refactor animation again
